cam_generation/requirements.txt
fastapi
uvicorn
openai
python-docx
docx2pdf
pydantic
python-multipart
requests
python-dotenv